refactor(benchmarks-website): split api.rs into api/ module
Split the 1702-line server/src/api.rs into a directory module with one
topical file per concern: dto (wire-shape structs), window (CommitWindow
+ ChartQuery), groups (group discovery passes), summary (v2-compatible
rollups), charts (per-fact-table chart payloads + SeriesAccumulator),
and filter (chip universe). The api/mod.rs surface re-exports the
existing public names so callers in html.rs and tests don't have to
update imports.

Behaviour preserved: every wire shape, SQL query, and snapshot test
matches byte-for-byte. The only structural change is a small helper
push_window_limit in charts.rs that pulls the repeated
CommitWindow::limit_param push out of the five chart collectors.
